requirements.txt
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorcet==2.0.2
cryptography==3.3.1
cycler==0.10.0
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Cors==3.0.9
Flask-Sockets==0.2.1
gevent==20.12.1
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
param==1.10.0
Pillow==8.0.1
pycparser==2.20
pyct==0.4.8
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
runway-python==0.6.1
scipy==1.5.4
six==1.15.0
torch==1.7.1
torch-dreams==1.1.0
torchvision==0.8.2
tqdm==4.55.0
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0